Fix review remarks
